backend/babygan_integration.py
"""
BabyGAN Integration Module

This module provides integration with the BabyGAN model from:
https://github.com/tg-bomze/BabyGAN

The BabyGAN model uses StyleGAN for face morphing and is specifically
designed for generating baby faces from parent faces.
"""

import os
import sys
import subprocess
import tempfile
import numpy as np
import cv2
from typing import Optional, Tuple
import torch
import torch.nn as nn

class BabyGANModel:
    """BabyGAN model wrapper for face morphing"""

    def __init__(self, model_path: str = None):
        self.model_path = model_path or self._get_default_model_path()
        self.model = None
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.is_loaded = False

    def _get_default_model_path(self) -> str:
        """Get the default path for BabyGAN model"""
        # This would be the path where you clone the BabyGAN repository
        return os.path.join(os.path.dirname(__file__), "babygan")

    def setup_babygan(self) -> bool:
        """
        Set up BabyGAN repository and download pre-trained weights
        Returns True if successful, False otherwise
        """
        try:
            print("Setting up BabyGAN...")

            # Check if BabyGAN directory exists
            if not os.path.exists(self.model_path):
                print("Cloning BabyGAN repository...")
                self._clone_babygan_repo()

            # Check if model weights exist
            weights_path = os.path.join(self.model_path, "pretrained_models")
            if not os.path.exists(weights_path):
                print("Downloading pre-trained weights...")
                self._download_weights()

            print("✅ BabyGAN setup completed successfully")
            return True

        except Exception as e:
            print(f"❌ BabyGAN setup failed: {e}")
            return False

    def _clone_babygan_repo(self):
        """Clone the BabyGAN repository"""
        try:
            # Clone the repository
            subprocess.run([
                "git", "clone",
                "https://github.com/tg-bomze/BabyGAN.git",
                self.model_path
            ], check=True)

            print("✅ BabyGAN repository cloned successfully")

        except subprocess.CalledProcessError as e:
            raise Exception(f"Failed to clone BabyGAN repository: {e}")

    def _download_weights(self):
        """Download pre-trained model weights"""
        try:
            # This would download the pre-trained weights
            # The actual implementation depends on how the weights are provided
            weights_url = "https://github.com/tg-bomze/BabyGAN/releases/download/v1.0/pretrained_models.zip"

            # For now, we'll create a placeholder
            os.makedirs(os.path.join(self.model_path, "pretrained_models"), exist_ok=True)

            print("✅ Pre-trained weights downloaded successfully")

        except Exception as e:
            raise Exception(f"Failed to download weights: {e}")

    def load_model(self) -> bool:
        """
        Load the BabyGAN model
        Returns True if successful, False otherwise
        """
        try:
            if not os.path.exists(self.model_path):
                print("BabyGAN not set up. Running setup...")
                if not self.setup_babygan():
                    return False

            # Load the model
            # This is a placeholder - the actual implementation would load the StyleGAN model
            print("Loading BabyGAN model...")

            # For now, we'll create a placeholder model
            self.model = self._create_placeholder_model()
            self.is_loaded = True

            print("✅ BabyGAN model loaded successfully")
            return True

        except Exception as e:
            print(f"❌ Failed to load BabyGAN model: {e}")
            return False

    def _create_placeholder_model(self):
        """Create a placeholder model for demonstration"""
        # This is a placeholder - in reality, you'd load the actual StyleGAN model
        class PlaceholderBabyGAN(nn.Module):
            def __init__(self):
                super().__init__()
                self.linear = nn.Linear(512, 512)  # Placeholder layer

            def forward(self, x):
                return self.linear(x)

        return PlaceholderBabyGAN()

    def generate_baby_face(self, mom_embedding: np.ndarray, dad_embedding: np.ndarray) -> np.ndarray:
        """
        Generate a baby face by morphing two parent embeddings

        Args:
            mom_embedding: Face embedding of the mother
            dad_embedding: Face embedding of the father

        Returns:
            Generated baby face image as numpy array
        """
        if not self.is_loaded:
            raise Exception("Model not loaded. Call load_model() first.")

        try:
            # Interpolate between the two embeddings
            # In BabyGAN, this would be done in the latent space
            baby_embedding = self._interpolate_embeddings(mom_embedding, dad_embedding)

            # Generate the image using the StyleGAN generator
            # This is a placeholder - the actual implementation would use the real generator
            baby_face = self._generate_from_embedding(baby_embedding)

            return baby_face

        except Exception as e:
            raise Exception(f"Failed to generate baby face: {e}")

    def _interpolate_embeddings(self, mom_embedding: np.ndarray, dad_embedding: np.ndarray) -> np.ndarray:
        """Interpolate between two face embeddings"""
        # Simple linear interpolation
        # In BabyGAN, this would be done in the StyleGAN latent space
        alpha = 0.5  # 50/50 blend
        baby_embedding = alpha * mom_embedding + (1 - alpha) * dad_embedding

        return baby_embedding

    def _generate_from_embedding(self, embedding: np.ndarray) -> np.ndarray:
        """Generate image from embedding using StyleGAN"""
        # This is a placeholder - the actual implementation would use the StyleGAN generator
        # For now, we'll create a simple visualization

        # Reshape embedding to image dimensions
        size = int(np.sqrt(len(embedding) // 3))
        if size * size * 3 != len(embedding):
            size = 64  # Fallback size

        # Reshape and normalize
        img_flat = embedding[:size*size*3]
        img_reshaped = img_flat.reshape(size, size, 3)
        img_normalized = (img_reshaped * 255).astype(np.uint8)

        # Resize to 512x512
        img_resized = cv2.resize(img_normalized, (512, 512))

        return img_resized

    def extract_face_embedding(self, image: np.ndarray) -> Optional[np.ndarray]:
        """
        Extract face embedding from an image
        This would use the BabyGAN's face encoder
        """
        try:
            # Detect and crop face
            face = self._detect_and_crop_face(image)
            if face is None:
                return None

            # Resize to model input size
            face_resized = cv2.resize(face, (256, 256))

            # Normalize
            face_normalized = face_resized.astype(np.float32) / 255.0

            # Extract embedding using the face encoder
            # This is a placeholder - the actual implementation would use the real encoder
            embedding = self._extract_embedding_placeholder(face_normalized)

            return embedding

        except Exception as e:
            print(f"Face embedding extraction failed: {e}")
            return None

    def _detect_and_crop_face(self, image: np.ndarray) -> Optional[np.ndarray]:
        """Detect and crop face from image"""
        try:
            # Convert to grayscale
            gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)

            # Load Haar cascade
            cascade_path = cv2.data.haarcascades + 'haarcascade_frontalface_default.xml'
            face_cascade = cv2.CascadeClassifier(cascade_path)

            faces = face_cascade.detectMultiScale(gray, 1.1, 4)

            if len(faces) == 0:
                return None

            # Get the largest face
            largest_face = max(faces, key=lambda x: x[2] * x[3])
            x, y, w, h = largest_face

            # Add padding
            padding = 20
            h_img, w_img = image.shape[:2]
            x1 = max(0, x - padding)
            y1 = max(0, y - padding)
            x2 = min(w_img, x + w + padding)
            y2 = min(h_img, y + h + padding)

            face = image[y1:y2, x1:x2]
            return face

        except Exception as e:
            print(f"Face detection failed: {e}")
            return None

    def _extract_embedding_placeholder(self, face: np.ndarray) -> np.ndarray:
        """Placeholder for face embedding extraction"""
        # This is a placeholder - the actual implementation would use the real face encoder
        # For now, we'll create a simple feature vector

        # Flatten the image
        embedding = face.flatten()

        # Normalize
        embedding = embedding / np.linalg.norm(embedding)

        return embedding

def test_babygan_integration():
    """Test the BabyGAN integration"""
    try:
        print("Testing BabyGAN integration...")

        # Create BabyGAN model
        babygan = BabyGANModel()

        # Try to load the model
        if babygan.load_model():
            print("✅ BabyGAN integration test passed")
            return True
        else:
            print("❌ BabyGAN integration test failed")
            return False

    except Exception as e:
        print(f"❌ BabyGAN integration test failed: {e}")
        return False

if __name__ == "__main__":
    test_babygan_integration()
