backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
# insightface==0.7.3  # Commented out due to macOS compilation issues
onnxruntime==1.16.3
scipy==1.11.4
scikit-image==0.22.0
aiofiles==23.2.1
python-dotenv==1.0.0
mediapipe==0.10.7
dlib==19.24.2
